+++ tests/test_coding/test_sequences.py
@@ -1,8 +1,8 @@
 import itertools
 import unittest
 
-from satella.coding.sequences import choose, infinite_counter, take_n, is_instance, is_last, \
-    add_next, half_cartesian, skip_first, zip_shifted, stop_after, group_quantity, \
+from satella.coding.sequences import choose, choose_one, infinite_counter, take_n, is_instance, \
+    is_last, add_next, half_cartesian, skip_first, zip_shifted, stop_after, group_quantity, \
     iter_dict_of_list, shift, other_sequence_no_longer_than, count, even, odd, Multirun, n_th
 
 
@@ -119,10 +119,11 @@ class TestSequences(unittest.TestCase):
             a = next(p)
             self.assertEqual(a, i)
 
-    def test_choose(self):
-        self.assertEqual(choose(lambda x: x == 2, [1, 2, 3, 4, 5]), 2)
-        self.assertRaises(ValueError, lambda: choose(lambda x: x % 2 == 0, [1, 2, 3, 4, 5]))
-        self.assertRaises(ValueError, lambda: choose(lambda x: x == 0, [1, 2, 3, 4, 5]))
+    def test_choose_one(self):
+        self.assertEqual(choose_one(lambda x: x == 2, [1, 2, 3, 4, 5]), 2)
+        self.assertEqual(choose(lambda x: x % 2 == 0, [1, 2, 3, 4, 5]), 2)
+        self.assertRaises(ValueError, lambda: choose_one(lambda x: x % 2 == 0, [1, 2, 3, 4, 5]))
+        self.assertRaises(ValueError, lambda: choose_one(lambda x: x == 0, [1, 2, 3, 4, 5]))
 
     def test_is_instance(self):
         objects = [object(), object(), [], [], object()]