+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 # v2.8.20
 
+* optimization in `CacheDict`