+++ CHANGELOG.md
@@ -1,3 +1,6 @@
-# v2.8.20
+# v2.9
 
 * optimization in `CacheDict`
+* changed the logic in choose
+* changed the semantics for `choose`
+* added `choose_one`